# public/swagger.yaml
openapi: 3.0.0
info:
  title: 'Mente Livre API'
  description: 'API REST para el sistema de gestión de terapeutas y citas de Mente Livre'
  contact:
    name: 'Mente Livre'
    email: mentelivre.pe@gmail.com
  version: 1.0.0
servers:
  -
    url: 'https://backend.mentelivre.org/'
    description: 'Servidor de Desarrollo Local'
paths:
...
  '/therapists/{therapistId}/pricing':
    get:
      tags:
        - 'Therapist Pricing'
      summary: 'Obtener precios de un terapeuta'
      operationId: getTherapistPricing
      parameters:
        -
          name: therapistId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Lista de precios del terapeuta'
    post:
      tags:
        - 'Therapist Pricing'
      summary: 'Crear precio para un terapeuta'
      operationId: createTherapistPricing
      parameters:
        -
          name: therapistId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: 'Precio creado exitosamente'
        '400':
          description: 'Datos inválidos'
      security:
        -
          bearerAuth: []
  '/therapist-pricing/{id}':
    get:
      tags:
        - 'Therapist Pricing'
      summary: 'Obtener precio por ID'
      operationId: getTherapistPricingById
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Precio encontrado'
        '404':
          description: 'Precio no encontrado'
    put:
      tags:
        - 'Therapist Pricing'
      summary: 'Actualizar precio'
      operationId: updateTherapistPricing
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Precio actualizado exitosamente'
        '404':
          description: 'Precio no encontrado'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - 'Therapist Pricing'
      summary: 'Eliminar precio'
      operationId: deleteTherapistPricing
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Precio eliminado exitosamente'
        '404':
          description: 'Precio no encontrado'
      security:
        -
          bearerAuth: []
  '/therapists/{therapistId}/pricing/batch':
    put:
      tags:
        - 'Therapist Pricing'
      summary: 'Actualizar múltiples precios de un terapeuta'
      operationId: updateTherapistPricingBatch
      parameters:
        -
          name: therapistId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Precios actualizados exitosamente'
      security:
        -
          bearerAuth: []
...
  '/users/{userId}/roles':
    get:
      tags:
        - 'User Roles'
      summary: 'Obtener roles de un usuario'
      operationId: getUserRoles
      parameters:
        -
          name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Lista de roles del usuario'
      security:
        -
          bearerAuth: []
    post:
      tags:
        - 'User Roles'
      summary: 'Asignar rol a un usuario'
      operationId: assignUserRole
      parameters:
        -
          name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - role_name
              properties:
                role_name:
                  type: string
                  example: therapist
              type: object
      responses:
        '200':
          description: 'Rol asignado exitosamente'
        '400':
          description: 'Error al asignar rol'
      security:
        -
          bearerAuth: []
  '/users/{userId}/roles/{roleName}':
    delete:
      tags:
        - 'User Roles'
      summary: 'Remover rol de un usuario'
      operationId: removeUserRole
      parameters:
        -
          name: userId
          in: path
          required: true
          schema:
            type: string
        -
          name: roleName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Rol removido exitosamente'
        '400':
          description: 'Error al remover rol'
      security:
        -
          bearerAuth: []
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Token JWT de autenticación. Incluir en el header: Authorization: Bearer {token}'
      bearerFormat: JWT
      scheme: bearer
...